chore: remove the _universe suffix from autoware_mission_planner
