+++ planning/autoware_mission_planner/plugins/plugin_description.xml
@@ -1,0 +1,3 @@
+<library path="autoware_mission_planner_lanelet2_plugins">
+  <class type="autoware::mission_planner::lanelet2::DefaultPlanner" base_class_type="autoware::mission_planner::PlannerPlugin"/>
+</library>
