+++ planning/autoware_mission_planner/schema/mission_planner.schema.json
