--- planning/autoware_mission_planner_universe/package.xml
@@ -1,44 +1,0 @@
-<?xml version="1.0"?>
-<?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
-<package format="3">
-  <name>autoware_mission_planner_universe</name>
-  <version>0.0.0</version>
-  <description>The autoware_mission_planner_universe package</description>
-  <maintainer email="isamu.takagi@tier4.jp">Takagi, Isamu</maintainer>
-  <maintainer email="kosuke.takeuchi@tier4.jp">Kosuke Takeuchi</maintainer>
-  <maintainer email="ryohsuke.mitsudome@tier4.jp">Ryohsuke Mitsudome</maintainer>
-  <maintainer email="takamasa.horibe@tier4.jp">Takamasa Horibe</maintainer>
-  <maintainer email="takayuki.murooka@tier4.jp">Takayuki Murooka</maintainer>
-  <maintainer email="mamoru.sobue@tier4.jp">Mamoru Sobue</maintainer>
-  <license>Apache License 2.0</license>
-
-  <author email="ryohsuke.mitsudome@tier4.jp">Ryohsuke Mitsudome</author>
-
-  <buildtool_depend>ament_cmake_auto</buildtool_depend>
-  <buildtool_depend>autoware_cmake</buildtool_depend>
-
-  <depend>autoware_adapi_v1_msgs</depend>
-  <depend>autoware_lanelet2_extension</depend>
-  <depend>autoware_map_msgs</depend>
-  <depend>autoware_motion_utils</depend>
-  <depend>autoware_planning_msgs</depend>
-  <depend>autoware_route_handler</depend>
-  <depend>autoware_utils</depend>
-  <depend>autoware_vehicle_info_utils</depend>
-  <depend>geometry_msgs</depend>
-  <depend>pluginlib</depend>
-  <depend>rclcpp</depend>
-  <depend>rclcpp_components</depend>
-  <depend>tf2_geometry_msgs</depend>
-  <depend>tf2_ros</depend>
-  <depend>tier4_planning_msgs</depend>
-
-  <test_depend>ament_cmake_ros</test_depend>
-  <test_depend>ament_lint_auto</test_depend>
-  <test_depend>autoware_lint_common</test_depend>
-  <test_depend>autoware_test_utils</test_depend>
-
-  <export>
-    <build_type>ament_cmake</build_type>
-  </export>
-</package>
