--- planning/autoware_mission_planner_universe/plugins/plugin_description.xml
@@ -1,3 +1,0 @@
-<library path="autoware_mission_planner_universe_lanelet2_plugins">
-  <class type="autoware::mission_planner_universe::lanelet2::DefaultPlanner" base_class_type="autoware::mission_planner_universe::PlannerPlugin"/>
-</library>
